hook in the extern runtime parameters.
Now if you put an _parameters file in any of the EOS, network, or
other microphysics directories (anything in EXTERN_CORE in Make.Castro)
then it will be added to extern_probin_module.  This change now actually
calls the runtime_init() routine in extern.f90 to initialize the
runtime parameters.  You can override the default values by adding
an &extern namelist section to your probin file -- if you don't do
this then we just stay with the defaults.