Free diffusion source terms from reliance on add_ext_src
The diffusion source terms are now handled with their own MultiFabs
in a way that is analogous to how the ext_src_old and ext_src_new
are handled. The subtlety comes with the numerous calls to computeTemp
that are sprinkled throughout Castro_advance.cpp. These can change
behavior significantly if not done in the same order. We therefore
insert a computeTemp call at the beginning of the new-time diffusion
source term update, which was happening before if we had add_ext_src
== 1 (which was necessary to do diffusion properly).

Though note that at present the new-time diffusion source terms are
equivalent to the old ones, so in practice the computeTemp call was
the only difference.

Fixes #1