Add an option to limit fluxes so that rho > small_dens
The new option is castro.limit_fluxes_on_small_dens. By setting it to 1,
a new piece of code is called after the fluxes are constructed that checks
if the sum of the fluxes leaving that zone would cause the density to
drop below the small_dens limit. If so, it limits those fluxes (and only
those fluxes, not the fluxes that would increase the density) such that
the resulting density is above the limit. This is intended to provide an
outlet that permits the avoidance of enforce_minimum_density (though the
latter is still used even when this option is on). It is not currently
fully conservative, though, since there can be differences in fluxes
constructed at box edges.

At present it is only implemented in 3D.