Create a flag that allows Castro to tell BoxLib to stop the run
The signalStopJob flag, if true, will set okToContinue() to false,
telling BoxLib to stop at the end of this step and write a checkpoint.
It starts the run as false and can be triggered using the stopJob() function.
This works in addition to the existing functionality that discontinues
the run if the timestep gets smaller than dt_cutoff.