Have meth_params_module read in C++ parameters directly
This removes the previous dependence on the C++ ParmParse querying
by having the queries occur in the meth_params_module. Thus
the results shouldn't change, but this cleans up the interfaces a bit
and means that we can read in the Castro parameters from the inputs file
without needing to go through the Castro::variableSetUp() routine.