Group the new-time source term construction/application
This changes the order of operations a bit; in particular, we used to
apply the source terms one by one and then for most of them do a
computeTemp call in between, so that the order in which we applied
the various source terms mattered. Now every new-time source term
only sees the state updated by the hydro, so it is order-independent.