change the strategy for buildInfo.cpp/.o
previously, buildInfo.cpp was written independently of the compilation
of it into buildInfo.o.  Now the creation and compilation are part
of a single target/rule.  The main consequence of this is that if
the build fails, then we always rebuild buildInfo.cpp.  We have to
force this in a slightly hackish way: there is a phony target
.FORCE that is used to force the rebuild