Move refluxes until after the hydro step on finest_level
Previously we would reflux in post_timestep after all levels finer
than the current level had proceeded. This had two related flaws.

First, the coarse timestep data was not updated as a result of
the fine grid timesteps until after the intermediate fine grid
steps had occurred, even though there was enough information at
the time to immediately update the state. This meant that in subsequent
interpolation steps when data was obtained from the coarse grid
to fill ghost zones or other data on the fine grid, we were not
using the most accurate version of the state available.

Second, the gravity sync solve at the end of the timestep was
also delayed relative to when the fluxes had actually occurred.
Now, with the gravity sync in particular, the source terms that we
calculated at the new time on the fine level were not correct.
This is more than just a re-statement of the first issue, because
previously we were attempting to correct for this with a sync source
applied after the sync solve had occurred. However, in general
with regridding it doesn't seem to be possible to calculate a
retroactive sync source that allows for the possibility that fine
grids may have appeared or disappeared over the course of the
coarse timestep, and even if you could do so, it would require
storage of the source terms you had applied as you went.

The drawback of this method is that it is more compute-intensive
because there's now one reflux and sync solve for every finest
level timestep. However, it is what is needed to ensure that
we have the most accurate value of the state data consistent
with the time integration strategy we use.