Do not save the comp_minus_level data if we're syncing
The comp_minus_level_phi data is useful for getting an answer for the
gravitational potential which is approximately correct, in that it provides
a lagged approximation to the composite solution. However, it is also incompatible
with the approach of doing a sync solve at the end of the timestep, because
the sync solve relies on the only difference between the coarse and fine
grids being the mismatch that develops on the coarse-fine interface. So,
a compromise solution is to use the comp_minus_level data when calculating the
forcing on the coarse grid (so that we get the most accurate force prior
to the sync solve occurring), but to then subtract it from the potential so
that the coming sync solve can proceed correctly.