Rewrite gravity source in terms of edge-centered gravity
When AMR is being used, edge-centered gravity is much more accurate
than the previously-used cell-centered potential, due to the fact that
the edge-centered gravity on refinement boundaries comes directly from
the multigrid, which uses a special stencil on the boundary to account
for the change in resolution. For non-Poisson gravity (e.g. monopole),
there is no change (we are constructing an edge-centered gravity by
averaging adjacent cell centers).