Add some protects against nonsense retry attempts
First, ensure that we're not about to do an integer overflow by requesting
more timesteps than integer counter can accept.

Second, check that we are not requesting more timesteps than a user-defined
maximum number. This is checked against the new parameter castro.retry_max_subcycles.
The parameter defaults to 1000.

Third, check that the subcycled timestep is not smaller than the user-defined
minimum timestep, castro.dt_cutoff.