Only swap time levels on the current level, not fine levels
This is in anticipation of the ability to add new levels at the end of
a timestep, not just at the beginning. This will mean that we need to
swap every time an advance is entered, because the level will not have
existed at the beginning of the coarse level timestep and so would
not have been swapped.

This commit should not change anything because we never need fine level
data before the fine level advance has begun.