Add post timestep regrid functionality
If the flag use_post_step_regrid is set to one, then a hook is
activated to check at the end of each timestep if any spatial
stability criteria have been violated. If so, we send a signal
to the Amr class that we want a regrid to occur immediately after
the end of the timestep. If this regrid results in a new fine
level being created, that fine level will immediately take its
timestep, rather than waiting for the next coarse step. In this way,
the fine grid will hopefully have enough resolution to not violate
the stability criterion.

At present the only stability criterion we have is the check against
numerically seeded detonations.

Note that for Poisson gravity, there is an added difficulty because
normally we do a multilevel solve to get the gravitational field in
sync after a regrid. We cannot do this here because the regrid occurs
after the coarse level has advanced, so the time levels are not in sync.
Consequently we simply rely on spatial interpolation to fill in the field.