Ensure that after a post-step regrid, the FluxRegister is updated
If a regrid occurs after a step, it will in generate create new grids on
the fine level (the level itself may even be newly created). This means that
when the fine level then takes a step immediately afterward, it needs to
throw out the previous coarse fluxes and re-initialize them. This relies on
the fact that we are saving the coarse fluxes.

As part of this change, the flux register CrseInit and FineAdd operations
were moved to their own Castro methods.