Use a MultiFab to store the fluxes during a timestep
The fluxes MultiFab holds this data normally when no subcycles occur.
But when subcycles occur, the fluxes MultiFab is intended to hold the
sum of the fluxes over all of the subcycles -- this is important for
getting refluxing right with AMR. But this introduces a problem because
the gravity source term references the fluxes MultiFab. A new MultiFab,
current_fluxes, has been introduced and it stores the hydro flux only
for a single timestep or subcycle.

Fixes #178