Do not swap gravity time levels on the first retry subcycle
The swapTimeLevels call for gravity copies the new data to the old data,
and sets the new data to a meaningless value. This makes sense for normal
evolution, but for subcycled evolution we must be careful because of the
way we use grad_phi. We do not simply recompute it from scratch during
the old-time level solve; rather, we use the saved value of it because
it contains data from the composite solve that is more accurate than the
level solve.

After the original level advance, during which we swapped the gravity
time levels, the old-time grad_phi is the correct composite data. So,
on the first subcycle of a retry iteration, we do not want to swap the
time levels -- everything is already OK (the new-time data does not have
the meaningless data in this case, but rather the new-time data from the
rejected advance, but this does not matter because the new-time data is
never re-used). On subsequent iterations, when we have advanced in time,
we want to swap the time levels as normal.