Add a README comment: all known bugs should have issues filed (#190)
Transparency about known bugs is important for helping users establish a judgment
about how reliable the code is. If they have done simulations using a given version
of the code they should be able to find out if there were any known issues relating
to the code modules they used. More broadly, setting an expectation of full transparency
about known issues helps establish a precedent in the community that it is acceptable
to make mistakes, and that being open about them improves confidence in the software
produced by the community as a whole.