Do not swap gravity time levels on the first retry subcycle (#189)
* Use a MultiFab to store the fluxes during a timestep

The fluxes MultiFab holds this data normally when no subcycles occur.
But when subcycles occur, the fluxes MultiFab is intended to hold the
sum of the fluxes over all of the subcycles -- this is important for
getting refluxing right with AMR. But this introduces a problem because
the gravity source term references the fluxes MultiFab. A new MultiFab,
current_fluxes, has been introduced and it stores the hydro flux only
for a single timestep or subcycle.

Fixes #178

* Revert "Use a MultiFab to store the fluxes during a timestep"

This reverts commit b72573539a23af48d6c900dcda5ff7917a390e14.

* Store the mass flux for use in evaluating the gravity source

* Change when mass_fluxes is allocated in a reflux

* Do not swap gravity time levels on the first retry subcycle

The swapTimeLevels call for gravity copies the new data to the old data,
and sets the new data to a meaningless value. This makes sense for normal
evolution, but for subcycled evolution we must be careful because of the
way we use grad_phi. We do not simply recompute it from scratch during
the old-time level solve; rather, we use the saved value of it because
it contains data from the composite solve that is more accurate than the
level solve.

After the original level advance, during which we swapped the gravity
time levels, the old-time grad_phi is the correct composite data. So,
on the first subcycle of a retry iteration, we do not want to swap the
time levels -- everything is already OK (the new-time data does not have
the meaningless data in this case, but rather the new-time data from the
rejected advance, but this does not matter because the new-time data is
never re-used). On subsequent iterations, when we have advanced in time,
we want to swap the time levels as normal.