Add options to limit the timestep to hit plot intervals exactly
The AMReX behavior for the input parameter plot_per, which is the simulation
time between taking plot files, is to generate a plot if we passed a multiple
of plot_per in the last timestep. If the timestep is of the same order as
plot_per, though, the resulting time of the plotfile can actually be significantly
different than a multiple of plot_per. If the timestep is significantly larger,
some multiples can be skipped altogether. The new parameter, castro.plot_per_is_exact,
enforces that a timestep that would pass a multiple of this interval is shortened
so that it hits this interval exactly. Then every plot file will be an exact multiple
of plot_per, and there will be a plotfile for every multiple of plot_per.

The same holds true for small plotfiles with the parameter castro.small_plot_per_is_exact.

To prevent very small timesteps, which can mess up the hydrodynamics, the shortened
timestep is limited to be no smaller than 10^-4 * dtEst, where dtEst is the limited dt
from all other factors (e.g. CFL limiting).