Remove an incorrect gravity time level swap during a retry
PR #189 did not correctly fix the issue described in #188 because it missed
an initial time level swap in the retry functionality. This is now fixed, so
that the first subcycle in a retry actually sees the same old gravity as the
rejected advance.

This changes results for problems that use retries with Poisson gravity.