When we force a post-timestep regrid, recalculate dt and subcycle if necessary
In this situation we either have a completely new level, or new zones on an existing
level, that was not factored in when we calculated the timestep at the beginning of the
coarse timestep. Consequently we can easily run into situations like CFL violations
or burning timestep constraint violations. In this case we calculate the timestep for
this level, and subcycle if necessary to complete the advance safely.