Take correct post-restart timestep if step was limited to hit plot interval
If a timestep was limited to hit a plot interval, a restart would see this shortened
dt when calculating the next dt, and would limit the timestep to change_max * dt.
This situation was worked around during normal evolution because we do not want to have
to catch back up to the longer timestep every time we stop to hit a plot interval.
This commit also ensures this behavior if we're restarting after a shortened timestep.

A new file is added to the checkpoint directory that stores the previous, longer timestep
in this situation. On a restart, we check if this file exists, and if so, read in that
longer timestep. This fixes the problem.

Fixes #242