Add parameter dxnuc_max to set an upper bound on dxnuc tagging
If dxnuc is used to refine zones above a certain burning instability threshold,
dxnuc_max can be used to set up an upper limit on the tagging. Zones that approach
order unity for dxnuc are essentially already ignited, so adding refinement to these
zones does not achieve the goal of suppressing numerically seeded ignitions, and
thus refinement for these zones may be unnecessary. Setting dxnuc_max allows the user
to further control the region of burning instability that is targeted for refinement.