Remove the parameter castro.keep_sources_until_end
This was added to assist with wdmerger due to the strategy of using
the source term data to calculate a relaxation step. This strategy
has been reworked and no longer needs the source term data, so this
parameter has been removed. Note that by default, the source term
data will be kept indefinitely anyway -- this is disabled only if
castro.update_sources_after_reflux = 0.