Move ctoprim to separate MFIter loop (#223)
The loop to do ctoprim is not slabbed like the PPM and Riemann solve are, so we get no real performance benefit from including it in the same MFIter loop. But we do take a performance hit because we are unnecessarily doing redundant work on ghost zones. This change removes the unnecessary ghost zone work, at the cost of allocating the whole array of primitive variables at one time.

* Fill q, qaux in a separate MFIter loop in MOL

* Fill q, qaux in a separate MFIter loop in CTU

* OMP the ctoprim loop

* Move compute_cfl up to C++

* Move ctoprim and the CFL check to separate functions

* Fix SDC compilation

* Apply the old sources to the hydro at the correct time

* Move Erborder and lamborder up to class scope