Fix the reflux source term correction after a retry (#266)
As #264 illustrates, the source term correction after a reflux is incorrect if a retry happened in the last timestep. This was because we were correcting based on the original old-time, but really we needed to be correcting based on the last subcycled timestep of the retry. This fixes #264 by saving that last timestep's data in the event of a retry, and temporarily swapping it in before re-calculating the new-time source terms after a reflux.

* First cut at fixing the retry issue with source terms

* Clean up the logic for saving the prev_state

* Revert some accidental test changes

* Fix logic

* Use a MF copy rather than pointer swap