Pass the state for the source terms as an argument to the source constructor
This generalizes the algorithm so that one can choose which state MultiFab
to send in for calculating the source terms. This is a step in the direction
of enabling source term calculations at arbitrary times, which is useful for
MOL, and also allows the old source terms to be either calculated with either
S_old or Sborder (which can be useful if trying to reconstruct the source
terms at a later time when Sborder is no longer available).

This changes results for diffusion if reactions are also enabled, since
diffusion was previously using the unreacted S_old state, not Sborder.