Rework the relaxation process using the approach of Rosswog et al. (2004)
In the past, the wdmerger relaxation used a source-term based approach similar
to that of Rosswog, Speith and Wynn (2004) and Dan et al. (2011), that balanced
the rotation force against the gravitational and hydrodynamic forces. This was
recently changed to a simpler method for various reasons, some of which were
related to Castro code changes (e.g. no longer explicitly storing the hydrodynamic
force) and the desire to make it easier to simulate only a half-domain (z > 0)
for symmetric problems; and, because the earlier implementation was found to not
be that robust anyway. The reason for this is probably that we were not really
doing what Rosswog et al. recommended. Rather, we were doing a related but different
approach where we updated the rotation force for the next timestep to balance
against what the gravitational and hydrodynamic forces were in this last
timestep. This meant that the rotation update was always lagging by a timestep,
which led to an ineffective method.

This commit actually does what Rosswog et al. recommends. At the end of the
timestep, we calculate what the rotation source term should have been in
that timestep to balance the gravitational and hydrodynamic forces, and subtract
off the rotation source term that was applied and replace it with the more
accurate one. This means the relaxation step does not lag. Since we no longer
store the source terms individually, this requires a little bit of programming
legwork to do, but we achieve the desired result.