change how we set variables to index state arrays (#275)
We now list variable indices in a plain text file _variables. These are used as the index keys into state
arrays both in C++ (for conserved state) and Fortran (for all the states). This is the first part of a clean-up of how we define variables. This addresses part of #267. At the moment, we run the script set_variables.py manually to write the Fortran routines that C++ include file. We also do some checking to ensure that we are synced up between C++ and Fortran.