Split the approximate state RIemann solvers so we can access the full interface state (#283)
The approximate state Riemann solvers are now split into two parts: the main solver routine now returns the full interface state (indexed as the primitive state with NQ components) and a separate routine computes the fluxes and the reduced set of Godunov state (with NGDNV components).  No changes to the test suite result.  This closes #279