Fix logic for dealing with CFL violations during retries
castro.hard_cfl_limit now has the behavior that it is ignored during the initial
advance if you're using a retry, and is enforced during the retries. Thus the default
behavior is now that when a CFL violation is detected and retries are enabled, a retry
will be taken, and if a subcycle in that retry violates CFL stability, the run aborts.

Fixes #334