Implement dynamic retry capability (#344)
Retries can now be done in a dynamic fashion. An advance is itself a subcycled advance always, and we keep subcycling until we are done with the step. By default we still do only a single timestep, but this helps in cases where we have to reject the timestep for (say) a CFL violation, and then during the retry the CFL criterion is violated again. In the past, we would simply have to abort the run if this happened. Now we can cut the timestep again and keep going.