Add tagging criteria for relative gradients
Tagging can now be done on relative gradients, in addition
to the existing capability for absolute gradients.
For example, tempgrad_rel is a relative gradient criterion
that will tag a zone with temperature T if any adjacent zone
has a temperature that is different by more than tempgrad_rel * T.
The tagging is enabled up to a given level with the parameter
max_tempgrad_rel_lev. The corresponding new tagging criteria
for other fields are named similarly.

Closes #354