Check whether there are any zones to burn before entering react_state().
If the user has set react_{rho,T}_{min,max} then we can save some time
before the burning has started in the simulation by checking to see if
any of the zones meet the minimum or maximum criteria for burning. If none
do, we can skip the call to react_state(). This ends up saving some cycles.