no slab version of CTU (#408)
Removes slabbed update from CTU that looped over the k index and for each k, did the entire 2D slice corresponding to that index. This approach was used for cache re-use benefits. This rewrite loses some of that benefit but makes the code more maintainable and better ready for a port to GPUs.