Ensure that the custom knapsack weights are always valid
This fixes #382, which reported the issue that the weights were being
set to the minimum of 1 and the number of RHS calls, which meant that
the weights were usually uniformly 1, so the custom knapsack was not
doing what we wanted. Also, in certain cases this could result in all
of the weights being zero, which caused a numerical issue and segfault
in the knapsack generation routine.