Enforce CFL constraint rigorously for old-time data
When the timestep check was originally written, we did not enforce violations
of the CFL timestep greater than 1 and less than 1 + 1 / CFL, because we found
that many small violations of the CFL criterion would pop up over a run, and
this caused many retries. But the retry_tolerance parameter was subsequently
added which allows for dealing with this more generally. Now, for safety, we
are re-asserting the full CFL constraint, and the user can override this with
a larger retry_tolerance parameter if they desire.

Also, we are now only applying the CFL constraint on the old-time data, not
the new-time data.