get rid of species sources by default (#519)
This change eliminates species sources by requiring us to set the parameter `PRIM_SPECIES_HAVE_SOURCES`. To support this the `set_variables.py` script was updated to now allow the "adds-to" field to have an optional if-def requirement.

This change also fixes 2 bugs. In `srctoprim`, we were converting the velocity sources twice. And in the radiation tracing, we were not adding species sources.

This closes #350