Remove clean_state from apply_source_to_state
apply_source_to_state() is currently called in three places in CTU:
after the old sources have been calculated, after the hydro source
has been calculated, and then after the new sources have been calculated.
For the first two, a call to clean_state is strictly unnecessary because
there's already a call to clean_state after the hydro is applied, and neither
the old source calculation nor the hydro calculation depends on the value
of S_new. So this should not change results, except to roundoff error (or
the iteration tolerance in the EOS). For the third, we can simply move the
clean_state call explicitly into the CTU advance. (We do the same for
the calls to apply_source_to_state in the reflux.)