Require clean_state to directly work on a state MultiFab
This change removes the indirect clean_state interface that took is_new
as an argument, and now clean_state must be directly handed the MultiFab
that it will operate on. This was done to partially address the concern
in #536 that expand_state is directly operating on S_old when it should
not be doing so. It also addresses the concern from #324 that the
behavior in expand_state was resulting in differences depending on whether
the state has ghost zones.

This change also removes the interface to clean_state that passes in an
old state for comparison. This was not necessary but it is nicer to
only have a single interface, and the logic there was a little complicated.
This choice necessitated a change in how we do enforce_minimum_density,
with two downstream consequences. First, density_reset_method == 3 no
longer exists -- since we don't have the old state to compare to, we cannot
reset to it. Second, because we don't have the old state, we have to change
the logic for the retry method based on negative densities. Now we look at
the change from the negative density to what it is reset to, rather than
the density it originally came from. This will change the outcome of that
method a bit, but qualitatively it does the same thing, since the hope is
that the density we reset to is at least in the same ballpark as the density
from the beginning of the step (we reset to a density similar to the
density in adjacent zones).