Refactor diffusion for GPU compatibility (#541)
This PR refactors the Fortran code for diffusion to be GPU compatible. To do so, the call to calculate the cell-centered conductivity term was split into a separate call from the routine that averages the cell-centered value to edges.

The MGutils files were made dimension agnostic and also set up for GPUing.

The diffusion_test problem was made compatible with running in 1D and 3D. This allows us to do performance testing with the latter.