Overhaul dimension agnostic BCs for GPU compatibility (#546)
This PR includes quite a few changes centered around making physical boundary conditions compatible for running on GPUs. Summary:

The dimension agnostic ca_*fill interfaces have been moved up to C++, leaving the actual routines that do the work (*fill) in Fortran.

The Fortran routines now take (lo, hi) as their first two arguments. They also now accept time by value rather than by reference.

The declarations have been split into separate headers (some in driver/, some in problems/).

We no longer do CUDA profiling on the initialization.

For CUDA, the number of registers when running on the GPU is capped to 128 when doing BCs. This is because amrex_filccn needs to call syncthreads() as it works to ensure synchronization, and because of the way it is designed, can need as many as 8 threads per dimension == 512 total, which is too many registers required if 255 registers per thread are allowed. This requires some custom GNU Make rules in Make.cuda_rules