fix thermo_src for MOL and SDC integration (#545)
This adds thermo_src (-pdivu U for rho e evolution) to source_flag() fucntion, to ensure that
we apply sources when using MOL and SDC integration.  It also ensures that CTU and the
simplified-SDC do not use this thermo source, since they discretize it directly in the conservative
update.

This also fixes a sign error on the source term itself, makes it work for non-Cartesian geom, and
disables update_sources_after_reflux for MOL and SDC integration.