Fix a CUDA synchronization problem in the symmetry/reflect BCs
For normal foextrap BCs, we fill the state from domlo(1) - NUM_GROW ... domlo(1) - 1
based on the state at domlo. Then we fill the domlo(2) - NUM_GROW ... domlo(2) - 1 values,
and the values in the corner require this to be done in order, which is why we synchronize
between the x, y, and z dimensions. This means that we need at least 5 zones per threadblock
in each dimension (5x5x5 = 125 in 3D) -- the 4 ghost zones and the first interior zone.

However for reflecting BCs, each ghost zone is based on its interior zone opposite to the
boundary at equal distance from the boundary. So the zone at -4 is based on zone 3, meaning
we need at least 8 zones per dimension. 8x8x8 = 512 threads per block, which is legal, and
should work since we cap registers for this kernel. But we were not enforcing this requirement -
the threadblock minimum size was assuming only enough threads to cover the `foextrap` case.
This has now been fixed: `hoextrap` and `reflect_even`/`reflect_odd` are now represented.