Add capability for GPU-compatible problem initialization
When we're running a problem on GPUs, we're typically only running 1 MPI rank per
GPU. On a system like Summit, this means 6 MPI ranks per node, each with a fairly
large box size (typically we'll want something like 128^3). But this single CPU
core isn't enough to do CPU-side operations quickly, so problem initialization
becomes a bottleneck (especially given that we don't currently have OpenMP support
when we're using CUDA, so we can't currently thread the initialization). This
commit adds the option for a problem to opt-in to a version of ca_initdata that
is GPU friendly. The interface is slightly different, and supports the GPU pragma.
The user has the responsibility to ensure that all of the variables accessed in
ca_initdata are GPU friendly, so this will not necessarily make sense for every problem.

This also fixes a bug introduced in commit ea93ddb in August 2015, where we were
giving incorrect bounds to ca_initdata after a restart for grown problems. Thus
this fixes #566.