Calculate offsets relative to problo, not xlo, in derives
This resolves an issue on the GPU: calculating offsets with respect to the
lo end of the box is not going to work, since we have lo == hi that matches
the thread index. Rather than try to get clever, we just calculate the grid
position using the index and the lo end of the domain using problo.