Fix a race condition in CTU hydro for 2D non-Cartesian
The Elixir for pradial was scoped inside the conditional that tested
on whether we're Cartesian, so the memory was eligible to be reclaimed
immediately. This then will cause an issue when there are multiple
boxes or multiple tiles. Now the Elixir is scoped at the MFIter level,
so the race condition is avoided.

Fixes #570