Decrease the amount of work done in the source_nonzero check (#571)
We only need to check the local tile, not the whole box. With this insight,
we can also now do this check on the GPU too, since the GPU will only be doing
the work of its current index +/- 2 zones, which gives a decent speedup there,
compared to doing the PPM reconstruction if the sources really are zero.