Implement HSE BC fill on GPUs
This required splitting out the HSE fill from the main hypfill/denfill
routines, because the HSE fill calls the EOS, which has not been
register-capped, so it needs to be done in a separate kernel launch.
The HSE fill is implemented without requiring any synchronizations:
a single thread does the whole column. So there is no need to require
a specific number of threads in the block, the same way we do with
the normal BC fill.