Improve ambient rotation velocity in inertial frame
The inertial frame setup for wdmerger forced the ambient material to
be stationary. This was done because rotating the ambient material at
the rigid body rotation rate would lead to unbounded velocities at
large distances from the center, which could cause undesired effects
such as limiting the hydrodynamic timestep. However, it also caused the
effect of the stars plowing through the stationary ambient material in
the inertial frame, which also can lead to small timesteps, especially
at low resolution.

A compromise is found between these two effects by applying the rigid
rotation rate in the region near the stars, but capping the rotation
rate as the edge of the domain is approached.