Riemann solver can not generate interface velocities > c
Castro is a Newtonian hydrodynamics code and has no knowledge of special
relativity. This means that spuriously large, unphysical velocities can
be generated. This commonly occurs in the wdmerger problem at poorly
resolved density gradients near the edge of the stars. Although a proper
solution to this problem involves adding SR to the code, a simple approach
to improving the situation is to enforce a speed limit in the Riemann solve.
By default this speed limit is c, but the user can lower it (for example,
to a value like 0.1c which is still in the relativistic regime where Castro
results should be distrusted) if they know the range of realistic values
taken by the velocity for their problem.