Updates to flame wave analysis scripts
Completely rewrote the front tracking script. The output and methods are now
configurable through command line arguments, and it should have a much
smaller memory footprint. The script for processing the data file was
updated accordingly.

Also updated the plot generator script with the bug fixes / new
additions that I've made over the past year or so.