move the code_comp initial model to probinit (#671)
This moves the generation of the initial model for code_comp to probinit, similar to how we do other problems. It is now managed by model_parser. The downside is that we need to specify the number of zones to use for the model generation, alteast until we deal with #669

* separate the initial model into a subroutine to make it easier to play with

* switch to model_parser to manage the model

* change coords

* fix allocates

* some conflicts