gpu compatible problems (#634)
This gets rid of the use of xlo in the Prob_nd.F90 files. This is a key step in the change to the GPU problem interface. This also converts a few problems over to the new auto prob_params
system.

Together these are enough to show how to convert a problem to GPU friendly.

Addresses #622