Simplify flux limiter to have single limiter per interface
This resolves the issue in #646. We were getting non-deterministic behavior
because we were mixing and matching updates from the left and right side of
the interface. By only applying the limiter to whichever side of the interface
would go negative, and thus only obtaining a single theta for the limiter,
we sidestep the issue and also simplify the code.