Fix ca_apply_metric for GPUs (#694)
ca_apply_metric was wrong on the GPU for two reasons. First, it was launching work over a cell-centered box but updating two nodal Fabs, so it was missing zones. Second, the memory access pattern is bad, since it is not strided in memory. Probably this is suboptimal for CPUs too, there's not enough work involved in saving r that it makes sense to invert the loop order.