Lengthen the last subcycle timestep to avoid very small dt
Due to roundoff issues, it is possible for the current algorithm to divide
the main timestep into N smaller dt's that, when summed up, so not exactly
equal the desired target time. We already have protection against the case
where the final time is slightly larger than the target time (in which case
we shortened that timestep) but we had no protection against the case where
we would end up slightly short of the target time, in which case we'd have
to take a very short last timestep. This PR resolves that issue by making
that last timestep slightly longer.